feat: move IdTabelCell and MetadataTableCell to components/table
