+++ src/app/(routes)/portfolios/portfolios-data-table.tsx
@@ -31,11 +31,11 @@ import { PaginationLimitField } from '@/components/form/pagination-limit-field'
 import { Pagination, PaginationProps } from '@/components/pagination'
 import { FormProvider, UseFormReturn } from 'react-hook-form'
 import { PaginationDto } from '@/core/application/dto/pagination-dto'
-import { IdTableCell } from '@/components/id-table-cell'
+import { IdTableCell } from '@/components/table/id-table-cell'
 import { PortfolioType } from '@/types/portfolio-type'
 import { PortfolioResponseDto } from '@/core/application/dto/portfolios-dto'
 import { PortfolioViewResponseDTO } from '@/core/application/dto/portfolio-view-dto'
-import { MetadataTableCell } from '@/components/metadata-table-cell'
+import { MetadataTableCell } from '@/components/table/metadata-table-cell'
 
 type PortfoliosDataTableProps = {
   portfolios: PaginationDto<PortfolioViewResponseDTO> | undefined
