+++ src/app/(routes)/segments/page.tsx
@@ -1,32 +1,12 @@
 'use client'
 
 import React, { useEffect, useState } from 'react'
-import { EntityBox } from '@/components/entity-box'
 import { Button } from '@/components/ui/button'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator
-} from '@/components/ui/dropdown-menu'
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableHeader,
-  TableRow
-} from '@/components/ui/table'
-import { DropdownMenuTrigger } from '@radix-ui/react-dropdown-menu'
-import { MoreVertical } from 'lucide-react'
 import { useIntl } from 'react-intl'
 import ConfirmationDialog from '@/components/confirmation-dialog'
 import { useDeleteSegment, useListSegments } from '@/client/segments'
 import { SegmentResponseDto } from '@/core/application/dto/segment-dto'
 import { useConfirmDialog } from '@/components/confirmation-dialog/use-confirm-dialog'
-import { EmptyResource } from '@/components/empty-resource'
-import { isNil } from 'lodash'
 import { useCreateUpdateSheet } from '@/components/sheet/use-create-update-sheet'
 import {
   getCoreRowModel,
@@ -34,15 +14,8 @@ import {
   useReactTable
 } from '@tanstack/react-table'
 import { useOrganization } from '@/context/organization-provider/organization-provider-client'
-import { useParams } from 'next/navigation'
-import { EntityDataTable } from '@/components/entity-data-table'
-import { Skeleton } from '@/components/ui/skeleton'
 import { useQueryParams } from '@/hooks/use-query-params'
-import { Pagination } from '@/components/pagination'
-import { PaginationLimitField } from '@/components/form/pagination-limit-field'
-import { IdTableCell } from '@/components/id-table-cell'
 import { SegmentsSheet } from './segments-sheet'
-import { MetadataTableCell } from '../ledgers/[id]/MetadataTableCell'
 import { getBreadcrumbPaths } from '@/components/breadcrumb/get-breadcrumb-paths'
 import { Breadcrumb } from '@/components/breadcrumb'
 import { PageHeader } from '@/components/page-header'
