+++ src/app/(routes)/segments/segments-data-table.tsx
@@ -29,8 +29,8 @@ import { FormProvider, UseFormReturn } from 'react-hook-form'
 import { EntityDataTable } from '@/components/entity-data-table'
 import { Pagination, PaginationProps } from '@/components/pagination'
 import { PaginationDto } from '@/core/application/dto/pagination-dto'
-import { IdTableCell } from '@/components/id-table-cell'
-import { MetadataTableCell } from '../ledgers/[id]/MetadataTableCell'
+import { IdTableCell } from '@/components/table/id-table-cell'
+import { MetadataTableCell } from '@/components/table/metadata-table-cell'
 import { SegmentResponseDto } from '@/core/application/dto/segment-dto'
 import { SegmentsSheet } from './segments-sheet'
 import { useOrganization } from '@/context/organization-provider/organization-provider-client'
